Agregado los policies para limitar edicion y borrado

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\AskQuestionRequest;
 use App\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -64,7 +65,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (Gate::allows('update-question', $question)){
+            return view('questions.edit', compact('question'));
+        }
+        abort(403, "Acceso No Permitido");
+
     }
 
     /**
@@ -76,8 +81,13 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $question->update($request->only('title', 'body'));
-        return redirect()->route('questions.index')->with('success', 'Tu pregunta ha sido actualizada');
+
+        if (Gate::allows('update-question', $question)){
+            $question->update($request->only('title', 'body'));
+            return redirect()->route('questions.index')->with('success', 'Tu pregunta ha sido actualizada');
+        }
+        abort(403, "Acceso No Permitido");
+
     }
 
     /**
@@ -88,7 +98,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
-        return redirect()->route('questions.index')->with('success', 'Tu pregunta ha sido eliminada');
+        if (Gate::allows('delete-question', $question)){
+            $question->delete();
+            return redirect()->route('questions.index')->with('success', 'Tu pregunta ha sido eliminada');
+        }
+        return "no estas autorizado para borrar esto";
     }
 }
